Module unicode::strUnstable

Structs

GraphemeIndices
External iterator for grapheme clusters and byte offsets.

Graphemes
External iterator for a string's grapheme clusters.

Utf16Encoder
Iterator adaptor for encoding chars to UTF-16.

Utf16Items
An iterator that decodes UTF-16 encoded codepoints from a vector of u16s.

Words
An iterator over the words of a string, separated by a sequence of whitespace

Enums

Utf16Item
The possibilities for values decoded from a u16 stream.

Traits

UnicodeStr
Methods for Unicode string slices

Functions

is_utf16
Determines if a vector of u16 contains valid UTF-16

utf16_items
Create an iterator over the UTF-16 encoded codepoints in v, returning invalid surrogates as LoneSurrogates.

utf8_char_width
Given a first byte, determine how many bytes are in this UTF-8 character